requirements.txt
asgiref==3.5.2
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cycler==0.11.0
Django==4.0.5
django-cors-headers==3.12.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
djangorestframework-api-key==2.2.0
emoji==1.7.0
fonttools==4.33.3
idna==3.3
joblib==1.1.0
kiwisolver==1.4.2
lxml==4.9.0
matplotlib==3.5.2
neattext==0.1.3
nltk==3.7
numpy==1.22.4
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
regex==2022.6.2
requests==2.27.1
scikit-learn==1.1.1
scipy==1.8.1
seaborn==0.11.2
six==1.16.0
sklearn==0.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
text2emotion==0.0.5
textblob==0.17.1
threadpoolctl==3.1.0
tqdm==4.64.0
tzdata==2022.1
urllib3==1.26.9
